ID Data on PlantHub & CSS
Data from firebase shows up om PlantHub screen

Most styling is in one JS file
