+++ screens/PlantProfileScreen/PlantProfileScreen.js
@@ -1,43 +1,22 @@
-import { View, SafeAreaView, Text, Image, Switch, StyleSheet, useWindowDimensions, ScrollView } from 'react-native';
+import { View, SafeAreaView, Text, Image, Switch, plantProfileStylesheet, useWindowDimensions, ScrollView } from 'react-native';
 import React, {useState} from 'react';
 import { Header } from 'react-native/Libraries/NewAppScreen';
+import { plantProfileStyles } from '../../components/Styles/Styling';
+
 
 
 const PlantProfileScreen = () => {
 
   return (
-    <SafeAreaView style={styles.container}>
-         <Text style = {styles.errorText}> 
+    <SafeAreaView style={plantProfileStyles.container}>
+         <Text style = {plantProfileStyles.errorText}> 
          Coming Soon!
          </Text>
     </SafeAreaView>
   );
 }
 
-const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#FFFFFF',
-    },
-    text: {
-        fontSize: 15,
-        fontWeight: 'bold',
-      paddingTop: 20,
-      marginRight: 145,
-      marginBottom: 5,
-    },
-  errorText: {
-    alignItems: "center",
-      alignSelf: "center",
-      marginTop: 20,
-      marginBottom: 10,
-      fontSize: 25,
-      fontWeight: "bold",
-      padding: 125,
-  },
 
-
-});
       
 
 export default PlantProfileScreen
